# specs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BursaPay Developer Gateway API",
    "description": "Official REST API v1 for the BursaPay developer payment infrastructure platform.",
    "version": "1.0.0",
    "contact": {
      "name": "BursaPay Developer Support",
      "email": "developers@bursapay.com",
      "url": "https://bursapay.com"
    }
  },
  "servers": [
    {
      "url": "https://api.bursapay.com/api/v1",
      "description": "Production & Sandbox Gateway (switch via API Key)"
    }
  ],
  "paths": {
...
    "/payments/initialize/": {
      "post": {
        "summary": "Initialize a payment",
        "description": "Create a checkout session and retrieve authorization URL or access code.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Unique UUID v4 for request deduplication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "email"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Payment amount in NGN",
                    "example": 15000.0
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "customer@example.com"
                  },
                  "currency": {
                    "type": "string",
                    "default": "NGN",
                    "example": "NGN"
                  },
                  "reference": {
                    "type": "string",
                    "example": "ORD-2026-90214"
                  },
                  "callback_url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://yourapp.com/checkout/callback"
                  },
                  "channels": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "card",
                      "bank_transfer",
                      "ussd"
                    ]
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/transfers/bulk/": {
      "post": {
        "summary": "Initiate bulk payouts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transfers": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bulk transfer queued"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "BursaPay API Key (bp_sec_...)"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  }
}